Create cochran.minergate.conf
Cochran is a better coin than it used to be. This is the first ForkNote coin to discourage users from investing any other currency into it, otherwise it will lose value. Bitcoin, Bytecoin or any other cryptocurrency is not meant to be bought or sold. It is a currency. USE, DON'T BUY/SELL.